EAM/backend/requirements.txt
Flask==3.1.0
Flask-CORS==5.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
SQLAlchemy==2.0.36
PyMySQL==1.1.1
cryptography==44.0.0
python-dotenv==1.0.1
Werkzeug==3.1.3
